Refactor of organizer_process_key.go
A much simpler refactor than Claude did in the other refactor branch.
This refactor did eliminate the ADD_CHANGE_FILTER mode which presented
the user with the container choices as opposed to just typing on the
command line but greatly simplifies the code and with tab completion
added, it should be more than adequate.  Lots of other small cleanup
of modes that are no longer in use.